requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
beautifulsoup4==4.12.2
selenium==4.15.2
openai==1.12.0
pandas==2.2.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
PyPDF2==3.0.1
pydantic==2.5.0
python-multipart==0.0.6
webdriver-manager==4.0.1
openai-whisper
ffmpeg